fix PAM Postgres queries

+++ src/models/src/entity/pam/group_user_links.rs
@@ -49,7 +49,7 @@ impl PamGroupUserLink {
         let res: Vec<Self> = if is_hiqlite() {
             DB::hql().query_map(sql, params!(gid)).await?
         } else {
-            DB::pg_query(sql, &[&gid], 4).await?
+            DB::pg_query(sql, &[&(gid as i64)], 4).await?
         };
 
         Ok(res)
@@ -61,7 +61,7 @@ impl PamGroupUserLink {
         let res: Vec<Self> = if is_hiqlite() {
             DB::hql().query_map(sql, params!(uid)).await?
         } else {
-            DB::pg_query(sql, &[&uid], 4).await?
+            DB::pg_query(sql, &[&(uid as i64)], 4).await?
         };
 
         Ok(res)
